pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "python-binance>=1.0.29",
    "requests>=2.32.4",
    "ta>=0.11.0",
]
